+++ rsky-pds/src/apis/com/atproto/sync/get_latest_commit.rs
@@ -51,7 +51,7 @@ pub async fn get_latest_commit(
         Ok(res) => Ok(Json(res)),
         Err(error) => {
             tracing::error!("@LOG: ERROR: {error}");
-            Err(ApiError::RuntimeError)
+            Err(ApiError::from(error))
         }
     }
 }