+++ rsky-pds/src/apis/com/atproto/sync/get_repo.rs
@@ -70,7 +70,7 @@ pub async fn get_repo(
         Ok(res) => Ok(BlockResponder(res)),
         Err(error) => {
             tracing::error!("@LOG: ERROR: {error}");
-            Err(ApiError::RuntimeError)
+            Err(ApiError::from(error))
         }
     }
 }