+++ rsky-pds/src/apis/com/atproto/sync/list_blobs.rs
@@ -78,7 +78,7 @@ pub async fn list_blobs(
         Ok(res) => Ok(Json(res)),
         Err(error) => {
             tracing::error!("@LOG: ERROR: {error}");
-            Err(ApiError::RuntimeError)
+            Err(ApiError::from(error))
         }
     }
 }