+++ rsky-pds/src/apis/com/atproto/sync/list_repos.rs
@@ -126,7 +126,7 @@ pub async fn list_repos(
         Ok(res) => Ok(Json(res)),
         Err(error) => {
             tracing::error!("@LOG: ERROR: {error}");
-            Err(ApiError::RuntimeError)
+            Err(ApiError::from(error))
         }
     }
 }